+++ lib/body.ml
@@ -77,7 +77,7 @@ module Reader = struct
       t.read_scheduled <- false;
       t.on_eof <- default_on_eof;
       t.on_read <- default_on_read;
-      let { Httpaf.IOVec.buffer; off; len } = iovec in
+      let { Httpun_types.IOVec.buffer; off; len } = iovec in
       Faraday.shift t.faraday len;
       on_read buffer ~off ~len;
       (* Application is done reading, we can give flow control tokens back to
@@ -160,8 +160,8 @@ module Writer = struct
     | `Yield | `Close -> 0
     | `Writev iovecs ->
       let buffered = t.buffered_bytes in
-      let iovecs = Httpaf.IOVec.shiftv iovecs !buffered in
-      let lengthv = Httpaf.IOVec.lengthv iovecs in
+      let iovecs = Httpun_types.IOVec.shiftv iovecs !buffered in
+      let lengthv = Httpun_types.IOVec.lengthv iovecs in
       let writev_len = if max_bytes < lengthv then max_bytes else lengthv in
       buffered := !buffered + writev_len;
       let frame_info = Writer.make_frame_info ~max_frame_size stream_id in