+++ lib/request.ml
@@ -33,7 +33,7 @@
  *---------------------------------------------------------------------------*)
 
 type t =
-  { meth : Httpaf.Method.t
+  { meth : Httpun_types.Method.t
   ; target : string
   ; scheme : string
   ; headers : Headers.t
@@ -49,7 +49,7 @@ let pp_hum fmt { meth; target; scheme; headers } =
   Format.fprintf
     fmt
     "((method \"%a\") (target %S) (scheme %S) (headers %a))"
-    Httpaf.Method.pp_hum
+    Httpun_types.Method.pp_hum
     meth
     target
     scheme